Updated worker-condition.go and tree-example.json and postman with new structure. Added more explaining string formats of condition and actions.

+++ work.go
@@ -15,7 +15,7 @@ type Action struct {
 	ValueQuoteMetric string
 	BaseCurrency     string
 	QuoteCurrency    string
-	Quantity         float32
+	Quantity         float64
 	Value            float64
 }
